requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

# Data
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
statsmodels==0.14.2

# Visualization
plotly==5.22.0
kaleido==0.2.1

# PDF Report
reportlab==4.2.2

# Streamlit Frontend
streamlit==1.37.0
httpx==0.27.0

# Optional Features
python-pptx==0.6.23
PyJWT==2.9.0

# Testing
pytest==8.3.2
